update when player touch hat/jet will skip other hat/jet

+++ monkey/Assets/Scripts/BulletBehaviour.cs
@@ -8,10 +8,10 @@ public class BulletBehaviour : MonoBehaviour
     // Start is called before the first frame update
     [SerializeField] private float normalBulletSpeed = 20f;
     [SerializeField] private float physicsBulletSpeed = 14.5f;
-    [SerializeField] private float physicsDamage = 2f;
-    [SerializeField] private float normalDamage= 1;
+    [SerializeField] private int physicsDamage = 2;
+    [SerializeField] private int normalDamage= 1;
     [SerializeField] private float DestroyTime = 2.5f;
-    private float damage;
+    private int damage;
     private Rigidbody2D rb;
     private AudioController audioController;
     public enum bulletType
